+++ src/components/ui/alert-dialog.tsx
@@ -1,9 +1,9 @@
+import * as AlertDialogPrimitive from '@radix-ui/react-alert-dialog';
 import type { ReactElement } from 'react';
 import * as React from 'react';
-import * as AlertDialogPrimitive from '@radix-ui/react-alert-dialog';
 
 import cn from '../../lib/utils';
-import { buttonVariants } from './button';
+import { buttonVariants } from '../../constants/buttons';
 
 const AlertDialog = AlertDialogPrimitive.Root;
 
@@ -92,14 +92,14 @@ AlertDialogCancel.displayName = AlertDialogPrimitive.Cancel.displayName;
 
 export {
   AlertDialog,
-  AlertDialogPortal,
-  AlertDialogOverlay,
-  AlertDialogTrigger,
-  AlertDialogContent,
-  AlertDialogHeader,
-  AlertDialogFooter,
-  AlertDialogTitle,
-  AlertDialogDescription,
   AlertDialogAction,
   AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogOverlay,
+  AlertDialogPortal,
+  AlertDialogTitle,
+  AlertDialogTrigger,
 };
